test: Run unit and integration tests across all built-in HTTP transports
