+++ tests/integration/test_apify_client.py
@@ -4,17 +4,13 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-import pytest
-
 from .._utils import maybe_await
-from .conftest import ALL_HTTP_CLIENT_CLASSES
 from apify_client._models import UserPrivateInfo, UserPublicInfo
 
 if TYPE_CHECKING:
     from apify_client import ApifyClient, ApifyClientAsync
 
 
-@pytest.mark.parametrize('http_client_classes', ALL_HTTP_CLIENT_CLASSES, indirect=True)
 async def test_apify_client(client: ApifyClient | ApifyClientAsync) -> None:
     """Test basic apify client functionality."""
     user_client = client.user('me')
