+++ tests/unit/test_client_errors.py
@@ -7,7 +7,6 @@ from unittest.mock import Mock
 import pytest
 from werkzeug import Response
 
-from apify_client import ApifyClient, ApifyClientAsync
 from apify_client.errors import (
     ApifyApiError,
     ConflictError,
@@ -25,7 +24,10 @@ if TYPE_CHECKING:
     from pytest_httpserver import HTTPServer
     from werkzeug import Request
 
+    from apify_client import ApifyClient, ApifyClientAsync
     from apify_client.http_clients import HttpClient, HttpClientAsync
+
+pytestmark = pytest.mark.usefixtures('http_client_classes')
 
 _TEST_PATH = '/errors'
 _EXPECTED_MESSAGE = 'some_message'
@@ -81,24 +83,6 @@ def streaming_handler(_request: Request) -> Response:
         status=403,
         mimetype='application/octet-stream',
         headers={'Content-Length': str(len(RAW_ERROR))},
-    )
-
-
-@pytest.fixture
-def sync_client(httpserver: HTTPServer, http_client_class: type[HttpClient]) -> ApifyClient:
-    return ApifyClient.with_custom_http_client(
-        token='test',
-        api_url=httpserver.url_for('/').removesuffix('/'),
-        http_client=http_client_class(),
-    )
-
-
-@pytest.fixture
-def async_client(httpserver: HTTPServer, http_client_async_class: type[HttpClientAsync]) -> ApifyClientAsync:
-    return ApifyClientAsync.with_custom_http_client(
-        token='test',
-        api_url=httpserver.url_for('/').removesuffix('/'),
-        http_client=http_client_async_class(),
     )
